+++ CHANGELOG.md
@@ -6,12 +6,17 @@
 |Author | `Nets eCom`
 |Prefix | `EASY-PS1.7`
 |Shop Version | `1.7+`
-|Version | `1.0.0`
+|Version | `1.1.0`
 |Guide | https://tech.nets.eu/shopmodules
 |Github | https://github.com/Nets-eCom/prestashop-PS1.7
 
 ## CHANGELOG
 
+### Version 1.1.0 - Released 2022-03-24
+* Fixed : Bugfixes in checkout section(Not getting country iso_code_3).
+		  Bugfixes for not getting consumer details in Nets portal.
+		  Bugfixes for not displaying live keys in admin configuration.
+* Docs: Updated license, changelog and readme files.
 
 ### Version 1.0.0 - Released 2022-01-28
 * New : Nets Easy plugin release with hosted and embedded payment page support.